Fix Wformat for stringtf and convert path buffers depending on PATH_MAX to std::string

+++ src/zm_logger.cpp
@@ -524,8 +524,8 @@ void Logger::logPrint(bool hex, const char *filepath, int line, int level, const
           "INSERT INTO `Logs` "
           "( `TimeKey`, `Component`, `ServerId`, `Pid`, `Level`, `Code`, `Message`, `File`, `Line` )"
           " VALUES "
-          "( %ld.%06ld, '%s', %d, %d, %d, '%s', '%s', '%s', %d )",
-          now_sec, now_frac.count(), mId.c_str(), staticConfig.SERVER_ID, tid, level, classString,
+          "( %ld.%06" PRIi64 ", '%s', %d, %d, %d, '%s', '%s', '%s', %d )",
+          now_sec, static_cast<int64>(now_frac.count()), mId.c_str(), staticConfig.SERVER_ID, tid, level, classString,
           escapedString.c_str(), file, line);
       dbQueue.push(std::move(sql_string));
     } else {
